# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/sessions/current": {
      "get": {
        "operationId": "SessionController_getCurrentSession",
        "parameters": [
          {
            "name": "cookie",
            "required": false,
            "in": "header",
            "description": "The browser's cookie header, forwarded by the web server.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCurrentSessionResponseDto_Output"
                }
              }
            }
          }
        },
        "summary": "Read who the caller is signed in as, without changing anything",
        "tags": [
          "identity"
        ]
      },
      "delete": {
        "operationId": "SessionController_signOut",
        "parameters": [
          {
            "name": "cookie",
            "required": false,
            "in": "header",
            "description": "The browser's cookie header, forwarded by the web server.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Sign out of the current session",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/sessions/current/refresh": {
      "post": {
        "operationId": "SessionController_refreshSession",
        "parameters": [
          {
            "name": "cookie",
            "required": false,
            "in": "header",
            "description": "The browser's cookie header, forwarded by the web server.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Push the current session forward and reissue its cookie",
        "tags": [
          "identity"
        ]
      }
    },
...
    "/v1/workspaces": {
      "post": {
        "operationId": "WorkspaceController_createWorkspace",
        "parameters": [
          {
            "name": "cookie",
            "required": false,
            "in": "header",
            "description": "The browser's cookie header, forwarded by the web server.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateWorkspaceResponseDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "REQUEST_INVALID, WORKSPACE_NAME_EMPTY or WORKSPACE_NAME_TOO_LONG"
          },
          "401": {
            "description": "NOT_SIGNED_IN"
          },
          "409": {
            "description": "PROFILE_REQUIRED or ALREADY_IN_A_WORKSPACE"
          }
        },
        "summary": "Create a Workspace, with the caller as its Owner",
        "tags": [
          "workspace"
        ]
      }
    },
...
    "/v1/workspaces/{workspaceId}/members": {
      "get": {
        "operationId": "WorkspaceMemberController_listWorkspaceMembers",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "cookie",
            "in": "header",
            "description": "The browser's cookie header, forwarded by the web server.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListWorkspaceMembersResponseDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "NOT_SIGNED_IN"
          },
          "404": {
            "description": "WORKSPACE_NOT_FOUND"
          }
        },
        "summary": "List the Members of a Workspace and their Roles",
        "tags": [
          "workspace"
        ]
      }
    },
    "/v1/workspaces/{workspaceId}/invitations": {
      "get": {
        "operationId": "InvitationController_listWorkspaceInvitations",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "cookie",
            "in": "header",
            "description": "The browser's cookie header, forwarded by the web server.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListWorkspaceInvitationsResponseDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "NOT_SIGNED_IN"
          },
          "403": {
            "description": "NOT_ALLOWED_TO_SEE_INVITATIONS"
          },
          "404": {
            "description": "WORKSPACE_NOT_FOUND"
          }
        },
        "summary": "List the Invitations of a Workspace still waiting on someone, pending or expired",
        "tags": [
          "workspace"
        ]
      },
      "post": {
        "operationId": "InvitationController_sendInvitations",
        "parameters": [
          {
            "name": "cookie",
            "required": false,
            "in": "header",
            "description": "The browser's cookie header, forwarded by the web server.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendInvitationsRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendInvitationsResponseDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "REQUEST_INVALID, NO_ADDRESSES_TO_INVITE or TOO_MANY_INVITATIONS"
          },
          "401": {
            "description": "NOT_SIGNED_IN"
          },
          "403": {
            "description": "NOT_ALLOWED_TO_INVITE"
          },
          "404": {
            "description": "WORKSPACE_NOT_FOUND"
          }
        },
        "summary": "Invite several people to a Workspace by email, at most 10 at a time",
        "tags": [
          "workspace"
        ]
      }
    }
  },
  "info": {
    "title": "Thrive API",
    "description": "Thrive — an employee feedback platform.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "GetCurrentSessionResponseDto_Output": {
        "type": "object",
        "properties": {
          "account": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "format": "email",
                "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
              },
              "name": {
                "type": "string"
              },
              "profile": {
                "type": "object",
                "properties": {
                  "fullName": {
                    "type": "string"
                  },
                  "displayName": {
                    "type": "string"
                  },
                  "slug": {
                    "type": "string"
                  }
                },
                "required": [
                  "fullName",
                  "displayName",
                  "slug"
                ],
                "additionalProperties": false,
                "nullable": true
              }
            },
            "required": [
              "email",
              "name",
              "profile"
            ],
            "additionalProperties": false,
            "nullable": true
          },
          "needsRefresh": {
            "type": "boolean"
          }
        },
        "required": [
          "account",
          "needsRefresh"
        ],
        "additionalProperties": false
      },
...
      "CreateWorkspaceRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "teamSize": {
            "type": "string",
            "enum": [
              "JUST_ME",
              "FROM_2_TO_10",
              "FROM_11_TO_50",
              "OVER_50",
              null
            ],
            "nullable": true
          }
        },
        "required": [
          "name"
        ],
        "additionalProperties": false
      },
      "CreateWorkspaceResponseDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          }
        },
        "required": [
          "id"
        ],
        "additionalProperties": false
      },
...
      "ListWorkspaceMembersResponseDto_Output": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "accountId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "email": {
                  "type": "string"
                },
                "fullName": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "role": {
                  "type": "string",
                  "enum": [
                    "OWNER",
                    "ADMIN",
                    "MEMBER"
                  ]
                }
              },
              "required": [
                "accountId",
                "email",
                "fullName",
                "role"
              ],
              "additionalProperties": false
            }
          },
          "total": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "page": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "pageSize": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ],
        "additionalProperties": false
      },
      "ListWorkspaceInvitationsResponseDto_Output": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "invitationId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "email": {
                  "type": "string"
                },
                "role": {
                  "type": "string",
                  "enum": [
                    "OWNER",
                    "ADMIN",
                    "MEMBER"
                  ]
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "PENDING",
                    "EXPIRED"
                  ]
                },
                "expiresAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                }
              },
              "required": [
                "invitationId",
                "email",
                "role",
                "status",
                "expiresAt"
              ],
              "additionalProperties": false
            }
          },
          "total": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "page": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "pageSize": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ],
        "additionalProperties": false
      },
      "SendInvitationsRequestDto": {
        "type": "object",
        "properties": {
          "emails": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "emails"
        ],
        "additionalProperties": false
      },
      "SendInvitationsResponseDto_Output": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string"
                    },
                    "outcome": {
                      "type": "string",
                      "enum": [
                        "invited"
                      ]
                    },
                    "invitationId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    }
                  },
                  "required": [
                    "email",
                    "outcome",
                    "invitationId",
                    "expiresAt"
                  ],
                  "additionalProperties": false
                },
                {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string"
                    },
                    "outcome": {
                      "type": "string",
                      "enum": [
                        "already_member",
                        "already_invited",
                        "failed"
                      ]
                    }
                  },
                  "required": [
                    "email",
                    "outcome"
                  ],
                  "additionalProperties": false
                }
              ]
            }
          }
        },
        "required": [
          "results"
        ],
        "additionalProperties": false
      }
    }
  }
}